Contains the Agile, Requirements, & Design HW2
The Word document called "Agile_Requirements" has the Agile question answers and the ones for Requirements. The "Habitat_Running" image has evidence that I could run Wildlife_Tracker as an Importable Module. The "Dependencies" PowerPoint has the dependencies between the modules as boxes and arrows with the specific methods. And the rest came as part of the Wildlife_Tracker folder with the methods and "innits" for each class within each module.